Add utility for determining http.Request IP address
The http.Request struct has a RemoteAddr field, but this simply holds
the address from which the request was received which could be a proxy
server. To determine the true source of the request, we need to look at
the address chain in the X-Forwarded-For header. If a publicly-routable
IP address cannot be found in the X-Forwarded-For or X-Real-IP header,
the RemoteAddr field will be used as a fallback.