Support for Install Script (#2)
The Agent prompts for input by reading stdin on register, however stdin
may be a bash script that is being piped from the a curl request. When
this is the case, the agent reads the entire script and tries to parse
it/use it as the user's first name. This causes a lot of problems.

To fix, only prompt if the user has not given one of the required pieces
of information in a command line argument. Use pflag for better flat
parsing. The curl script asks for this information and passes it to the
agent.